# backend springboot/src/main/java/com/attendance/backend/domain/enums/NotificationSourceType.java
package com.attendance.backend.domain.enums;

public enum NotificationSourceType {
    ABSENCE_REQUEST,
    ATTENDANCE_POLICY,
    ATTENDANCE_SESSION,
    ATTENDANCE_EVENT,
    FRAUD_INCIDENT,
    ADMIN_ACTION,
    SECURITY_EVENT
}