# backend springboot/src/main/java/com/attendance/backend/domain/enums/NotificationType.java
package com.attendance.backend.domain.enums;

public enum NotificationType {
    ABSENCE_REQUEST_CREATED,
    ABSENCE_REQUEST_APPROVED,
    ABSENCE_REQUEST_REJECTED,
    ABSENCE_REQUEST_CANCELLED,
    ABSENCE_REQUEST_REVERTED,
    ATTENDANCE_POLICY_WARNING,
    ATTENDANCE_POLICY_CRITICAL,
    FRAUD_ALERT_RISK,
    FRAUD_ALERT_CRITICAL
}