# backend springboot/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Attendance QR API
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /me/notifications:
    get:
      tags:
        - Notifications
      summary: List my notifications
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/NotificationType'
        - name: severity
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/NotificationSeverity'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageNotificationResponse'
        '400':
          $ref: '#/components/responses/Err400'
        '401':
          $ref: '#/components/responses/Err401'
  /me/notifications/unread-count:
      get:
        tags:
          - Notifications
        summary: Get my unread notification count
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/NotificationUnreadCountResponse'
          '401':
            $ref: '#/components/responses/Err401'
  /me/notifications/{notificationId}/read:
      post:
        tags:
          - Notifications
        summary: Mark one notification as read
        parameters:
          - name: notificationId
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/NotificationResponse'
          '401':
            $ref: '#/components/responses/Err401'
          '404':
            $ref: '#/components/responses/Err404'
  /me/notifications/read-all:
      post:
        tags:
          - Notifications
        summary: Mark all my notifications as read
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/NotificationReadAllResponse'
          '401':
            $ref: '#/components/responses/Err401'
...
  /groups:
    post:
      tags:
        - Groups
      summary: Create group (caller becomes OWNER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '409':
          $ref: '#/components/responses/Err409'
        '400':
          $ref: '#/components/responses/Err400'
    get:
      tags:
        - Groups
      summary: List my groups
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageGroupResponse'
...
  /groups/{groupId}/notifications:
    get:
      tags:
        - Notifications
      summary: List notifications of current user in a group
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageNotificationResponse'
        '401':
          $ref: '#/components/responses/Err401'
        '404':
          $ref: '#/components/responses/Err404'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
    Sort:
      name: sort
      in: query
      schema:
        type: string
        example: createdAt,desc
  responses:
    Err400:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            validation:
              value:
                status: 400
                code: VALIDATION_ERROR
                message: Invalid request
    Err401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            unauthorized:
              value:
                status: 401
                code: UNAUTHORIZED
                message: Missing/invalid token
...
    Err404:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            notFound:
              value:
                status: 404
                code: NOT_FOUND
                message: Resource not found
    Err409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            conflict:
              value:
                status: 409
                code: CONFLICT
                message: Conflict
...
  schemas:
    NotificationType:
      type: string
    NotificationSeverity:
        type: string
    NotificationSourceType:
        type: string
...
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
          nullable: true
        sessionId:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/NotificationType'
        title:
          type: string
        body:
          type: string
        payload:
          type: object
          nullable: true
          additionalProperties: true
        severity:
          $ref: '#/components/schemas/NotificationSeverity'
        sourceType:
          $ref: '#/components/schemas/NotificationSourceType'
        sourceRefId:
          type: string
          format: uuid
          nullable: true
        isRead:
          type: boolean
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    NotificationUnreadCountResponse:
      type: object
      required:
        - unreadCount
      properties:
        unreadCount:
          type: integer
          format: int64
    NotificationReadAllResponse:
      type: object
      required:
        - updatedCount
        - readAt
      properties:
        updatedCount:
          type: integer
          format: int32
        readAt:
          type: string
          format: date-time
    PageNotificationResponse:
      type: object
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        path:
          type: string
        traceId:
          type: string
      required:
        - timestamp
        - status
        - code
        - message
        - path
...
    ApprovalMode:
      type: string
      enum:
        - AUTO
        - MANUAL
    GroupStatus:
      type: string
      enum:
        - ACTIVE
        - ARCHIVED
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 150
        code:
          type: string
          maxLength: 20
        joinCode:
          type: string
          minLength: 6
          maxLength: 16
        description:
          type: string
          maxLength: 1000
        semester:
          type: string
          maxLength: 30
        room:
          type: string
          maxLength: 80
        approvalMode:
          $ref: '#/components/schemas/ApprovalMode'
        allowAutoJoinOnCheckin:
          type: boolean
          default: false
      required:
        - name
        - code
        - approvalMode
...
    GroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 150
        code:
          type: string
          maxLength: 20
        joinCode:
          type: string
          maxLength: 16
        description:
          type: string
          maxLength: 1000
        semester:
          type: string
          maxLength: 30
        room:
          type: string
          maxLength: 80
        approvalMode:
          $ref: '#/components/schemas/ApprovalMode'
        allowAutoJoinOnCheckin:
          type: boolean
        status:
          $ref: '#/components/schemas/GroupStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ownerUserId
        - name
        - code
        - joinCode
        - approvalMode
        - status
        - createdAt
        - updatedAt
...
    PageGroupResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GroupResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
...